% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{openai_chat_completion}
\alias{openai_chat_completion}
\title{Send a request to the OpenAI chat completion endpoint}
\usage{
openai_chat_completion(
  messages,
  model = "gpt-4o-2024-08-06",
  api_params = list(temperature = 0.8),
  api_key = get_openai_api_key()
)
}
\arguments{
\item{messages}{The conversation up to now. Make one from scratch with
\code{openai_create_conversation}.}

\item{model}{The model to use. Defaults to the "best/cheapest" choice at time of writing.}

\item{api_params}{All other API parameters are optional. Provide them inside  a named list and they will be sent
as-is.}

\item{api_key}{The OpenAI API key to use. Defaults to env var \code{OPENAI_API_KEY}.}
}
\value{
An \code{httr2} response object if there was a problem. A named list containing \code{role} and \code{content} elements
otherwise.
}
\description{
Send a request to the OpenAI chat completion endpoint
}
