% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{openai_continue_conversation}
\alias{openai_continue_conversation}
\title{Continue an existing OpenAI cat completion conversation}
\usage{
openai_continue_conversation(
  conversation,
  role = c("user", "system", "assistant"),
  content
)
}
\arguments{
\item{conversation}{The conversation so far. List of lists.}

\item{role}{The role for the message we are adding to the conversation.
One of \code{user}, \code{system} or \code{assistant}.}

\item{content}{The message to add to the conversation.}
}
\value{
A list of lists. The existing conversation with our message added.
}
\description{
Continue an existing OpenAI cat completion conversation
}
