% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{openai_create_conversation}
\alias{openai_create_conversation}
\title{Create an OpenAI chat completion conversation}
\usage{
openai_create_conversation(user_message, system_message = NULL)
}
\arguments{
\item{user_message}{A prompt in ChatGPT parlance.}

\item{system_message}{Instruction to guide the behavior and personality of the model throughout the conversation.
A default one will be provided free of charge.}
}
\value{
A list of lists. Each element of the outer list contains a role and a message.
}
\description{
Create an OpenAI chat completion conversation
}
